refactor: categoryHeader 카테고리 페이지 내부 컴포넌트 폴더로 경로 수정
